Surgically removed the graph view into its own page file as MainPage.jsx from App.jsx
